Started  a system to manage training plans, exercises and sessions.
Developed a method for creating custom exercises to companies.
Implemented methods for generating custom exercises and custom sets for trainers.
Missing in the method for create custom sets a way to create a set with a default name.
Added role validation to the AuthInterceptor for all methods in the TrainerController class.
Planned additional methods for the session system.